Introduce ALSAdaptor-sysfs plugin
This patch adds a new ambient light sensor adaptor. The adaptor simply
periodically reads a sysfs node that is supposed to hold a lux
reading in ascii form. The plugin was written for Nokia N900 device but the
sysyfs file path is configurable so this adaptor is easily re-used for other
devices that have a similar driver interface.