Proximityadaptor-evdev plugin.
This patch adds a proximityadaptor that uses event device interface of the
kernel. The plugin is written for Nokia N900 and it currently has only one
configuration option (name to match the correct evdev driver). However, the
plugin can easily be made more configurable if other devices that have
proximity sensors exposed through evdev interface must be supported and the
current implementation is not flexible enough.