[sensorfw] Don't build anything but the client library for Qt 5.
As Qt 5 is functionally crippled due to the state of contextkit meaning
virtually no plugins are operational / load, it makes no sense to try to port
the daemon to Qt 5 at this point.

Instead, leave the daemon Qt 4, and add a requirement on it from the Qt 5 client
library packaging. This also avoids problems with both packages providing the
same paths.

In the future, we can reverse this situation: have Qt 4 only build the client
library, and port the daemon to Qt 5 for good.

As a happy coincidence, this means that when you install both sensorfw and
sensorfw-qt5, you no longer end up having to flip a coin to determine which
daemon you get.